change: alteração no tipo do ID nas entidades do banco de dados

+++ src/main/java/com/srm/demo/adapters/db/repositories/EmprestimoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.srm.demo.adapters.db.entities.Emprestimo;
 
 @Repository
-public interface EmprestimoRepository extends JpaRepository<Emprestimo, Integer>{
+public interface EmprestimoRepository extends JpaRepository<Emprestimo, Long>{
     
 }
